Update script.js
Removed some websites that are either defunct or I'm uninterested in. Added some websites that I frequently go to or that I'm too lazy to type in every time. Added a new post number function for nhentai. Do !nh and then type the sacred number you're looking for, as well as a search for Tsumino using !t